# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Gin personal blog backend API.",
        "title": "MyBlog API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/tag/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员创建文章标签。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/tag"
                ],
                "summary": "创建标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建标签请求体。tag_name 示例：Gin；slug 示例：gin；sort 示例：10",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"id\\\":1},\\\"msg\\\":\\\"创建标签成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"参数格式错误\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"创建标签失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/tag/{id}/sort": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员更新标签排序值。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/tag"
                ],
                "summary": "更新标签排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "标签 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "标签排序请求体。sort 示例：20",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateTagSortRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"msg\\\":\\\"更新标签排序成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"标签ID不合法\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"更新标签排序失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/tag/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员更新标签启用状态。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/tag"
                ],
                "summary": "更新标签状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "标签 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "标签状态请求体。status 示例：1 表示启用，0 表示停用",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateTagStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"msg\\\":\\\"更新标签状态成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"标签ID不合法\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"更新标签状态失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/tags": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员获取全部标签列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/tag"
                ],
                "summary": "获取后台标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":[],\\\"msg\\\":\\\"获取标签列表成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"获取标签列表失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/tags": {
            "get": {
                "description": "获取启用状态的标签列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "获取前台标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "无业务参数",
                        "name": "unused",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":[],\\\"msg\\\":\\\"获取标签列表成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"获取标签列表失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreateTagRequest": {
            "type": "object",
            "required": [
                "slug",
                "sort",
                "tag_name"
            ],
            "properties": {
                "slug": {
                    "type": "string",
                    "example": "gin"
                },
                "sort": {
                    "type": "integer",
                    "example": 10
                },
                "tag_name": {
                    "type": "string",
                    "example": "Gin"
                }
            }
        },
...
        "controllers.UpdateTagSortRequest": {
            "type": "object",
            "required": [
                "sort"
            ],
            "properties": {
                "sort": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "controllers.UpdateTagStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}